First working version of TicTacToe
Run the code, an random player will be their turn first. They can then
enter the row and column they wish to move. It will print the board.
Then go to next player's turn. If there is a winner, it will print the
board, congratulate the player, and ask if you want to restart the game.
If the player does, the game will restart, else ends the current
session.